requirements.txt
fastapi
gradio
pandas
uvicorn
torch
torchvision
Pillow
numpy
python-multipart
opencv-python
requests
matplotlib
seaborn
